change file structure for maintainability
